conversion Message -> MSG moved to msgparser
