+++ msgparser/src/main/java/com/auxilii/msgparser/entries/GuidStreamEntry.java
@@ -1,4 +1,4 @@
-package net.sourceforge.MSGViewer.factory.msg.entries;
+package com.auxilii.msgparser.entries;
 
 import static com.auxilii.msgparser.Pid.PidTagGuidStream;
