+++ msgparser/src/main/java/com/auxilii/msgparser/entries/MessageClassEntry.java
@@ -1,9 +1,9 @@
-package net.sourceforge.MSGViewer.factory.msg.entries;
+package com.auxilii.msgparser.entries;
 
 import static com.auxilii.msgparser.Pid.PidTagMessageClass;
 
 public class MessageClassEntry extends StringUTF16SubstgEntry {
-    public static final String IPM_NOTE = "IPM.Note";
+    private static final String IPM_NOTE = "IPM.Note";
 
     public MessageClassEntry() {
         super(PidTagMessageClass, IPM_NOTE);
