+++ msgparser/src/main/java/com/auxilii/msgparser/entries/StringUTF16SubstgEntry.java
@@ -1,8 +1,8 @@
-package net.sourceforge.MSGViewer.factory.msg.entries;
+package com.auxilii.msgparser.entries;
 
 import com.auxilii.msgparser.Pid;
-import net.sourceforge.MSGViewer.factory.msg.properties.PropPtypString;
-import net.sourceforge.MSGViewer.factory.msg.properties.PropType;
+import com.auxilii.msgparser.properties.PropPtypString;
+import com.auxilii.msgparser.properties.PropType;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
